// apps/web/models/group/enums/group-permission.enum.ts
export enum GroupPermission {
  // Group
  MANAGE_GROUP = 'MANAGE_GROUP',
  MANAGE_MEMBERS = 'MANAGE_MEMBERS',
  UPDATE_GROUP = 'UPDATE_GROUP',

  // Posts
  APPROVE_POST = 'APPROVE_POST',
  DELETE_POST = 'DELETE_POST',

  //Members
  BAN_MEMBER = 'BAN_MEMBER',
  VIEW_REPORTS = 'VIEW_REPORTS',

  // Settings
  VIEW_SETTINGS = 'VIEW_SETTINGS',
  UPDATE_GROUP_SETTINGS = 'UPDATE_GROUP_SETTINGS',

  // Manage Join Requests
  MANAGE_JOIN_REQUESTS = 'MANAGE_JOIN_REQUESTS',

  // Events
  MANAGE_EVENTS = 'MANAGE_EVENTS',

  // Invitations
  INVITE_MEMBERS = 'INVITE_MEMBERS',
}